# view/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "",
      "version": "2",
      "title": "Store"
    },
    "host": "localhost:3030",
    "tags": [
      {
        "name": "hospital"
      }
    ],
    "schemes": [
      "http"
    ],
    "paths": {
      "/totalpatients": {
        "post": {
          "tags": [
            "hospital"
          ],
          "summary": "order by price (low-->high)",
          "description": "order by price (low-->high)",
          "operationId": "order by price (low-->high)",
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "Query Success",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 201
                  },
                  "message": {
                    "type": "string",
                    "example": {
                      "HNID": "0001",
                      "count": 2
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 400
                  },
                  "message": {
                    "type": "string",
                    "example": "Bad Request"
                  }
                }
              }
            },
            "500": {
              "description": "unAuthenticate",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 500
                  },
                  "message": {
                    "type": "string",
                    "example": "Fail to Connect Database"
                  }
                }
              }
            }
          }
        }
      },
      "/datapatients": {
        "post": {
          "tags": [
            "hospital"
          ],
          "summary": "order by price (high-->low)",
          "description": "order by price (high-->low)",
          "operationId": "order by price (high-->low)",
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "Query Success",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 201
                  },
                  "message": {
                    "type": "string",
                    "example": {
                      "HNID": "0001",
                      "Firstname": "name",
                      "Surname": "Surname"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 400
                  },
                  "message": {
                    "type": "string",
                    "example": "Bad Request"
                  }
                }
              }
            },
            "500": {
              "description": "unAuthenticate",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 500
                  },
                  "message": {
                    "type": "string",
                    "example": "Fail to Connect Database"
                  }
                }
              }
            }
          }
        }
      },
      "/rankpatients": {
        "post": {
          "tags": [
            "hospital"
          ],
          "summary": "orankpatients",
          "description": "rankpatients",
          "operationId": "rankpatients",
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "Query Success",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 201
                  },
                  "message": {
                    "type": "string",
                    "example": {
                      "HID": "0001",
                      "count": 2
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 400
                  },
                  "message": {
                    "type": "string",
                    "example": "Bad Request"
                  }
                }
              }
            },
            "500": {
              "description": "unAuthenticate",
              "schema": {
                "properties": {
                  "statusCode": {
                    "type": "string",
                    "example": 500
                  },
                  "message": {
                    "type": "string",
                    "example": "Fail to Connect Database"
                  }
                }
              }
            }
          }
        }
      }
    }
  }